requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0

attrs==25.3.0
bcrypt==4.0.1
beautifulsoup4==4.13.4
bitlyshortener==0.6.4
black==25.1.0
blinker==1.9.0
blis==1.3.0
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
cloudpathlib==0.21.1
cobble==0.1.4
colorama==0.4.6
comtypes==1.4.11
confection==0.1.5
cymem==2.0.11
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docx2pdf==0.1.8
ecdsa==0.19.1
email_validator==2.2.0
environs==14.2.0
et_xmlfile==2.0.0
fastapi==0.116.1
fastapi_cors==0.0.6
filelock==3.18.0
fire==0.7.0
flake8==7.3.0
Flask==2.3.3
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-WTF==1.1.1
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
googlemaps==4.10.0
gotrue==2.12.3
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.25
langchain-text-splitters==0.3.8
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.4.9
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.4.2
language_data==1.3.0
lxml==5.4.0
mammoth==1.9.1
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==4.0.0
markdown2==2.5.3
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.47.0
networkx==3.5
numpy==2.3.0
openai==1.97.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
pathspec==0.12.1
pdf2docx==0.5.8
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
portalocker==2.10.1
postgrest==1.1.1
preshed==3.0.10
prometheus-client==0.19.0
propcache==0.3.2
protobuf==6.31.1
psutil==5.9.6

pyarrow==20.0.0
pybreaker==1.0.2
pycodestyle==2.14.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.1
pyngrok==7.2.12
pytest==8.4.1
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-http-client==3.3.7
python-multipart==0.0.6
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.3
realtime==2.5.3
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rq==1.15.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sendgrid==6.10.0
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlite-fts4==1.0.3
sqlite-utils==3.38
sqlmodel==0.0.24
srsly==2.5.1
starkbank-ecdsa==2.2.0
starlette==0.46.2
storage3==0.12.0
streamlit==1.47.0
StrEnum==0.4.15
structlog==25.4.0

supafunc==0.10.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
twilio==9.6.5
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
Werkzeug==2.3.7
wheel==0.45.1
wrapt==1.17.3
WTForms==3.2.1
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
